whiskeyservice

+++ src/main/java/services/SneakerService.java
@@ -21,7 +21,7 @@ public class SneakerService {
         return sneaker;
     }
 
-    public Sneaker find(Long id) {
+    public Sneaker find(Integer id) {
         for (Sneaker sneaker : inventory) {
             if (sneaker.getId().equals(id)) {
                 return sneaker;
@@ -30,7 +30,7 @@ public class SneakerService {
         return null;
     }
 
-    public Sneaker update(Sneaker sneaker, Long id) {
+    public Sneaker update(Sneaker sneaker, Integer id) {
         for (Sneaker s : inventory) {
             if (s.getId().equals(id)) {
                 s.setName(sneaker.getName());
@@ -46,7 +46,7 @@ public class SneakerService {
         return null;
     }
 
-    public boolean delete(Long id) {
+    public boolean delete(Integer id) {
         for (Sneaker s : inventory) {
             if (s.getId().equals(id)) {
                 inventory.remove(s);
